Tile fixes. Added water layer. Organized tile folder assets.
Few minor changes to tiles for overlap design style. Added a water layer
to layer sorting between background and default, allowing the player to
be in front of water tiles. Organized the tile folder. Textures are in
the tile folder and asset tiles are placed into the asset tile folder.